extraction-service/requirements.txt
flask==2.3.3
pytesseract==0.3.10
Pillow==10.0.1
gunicorn==21.2.0